+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 CHANGES:
 
+1. Add `process_monver()` helper method to extract dictionary of hardware, firmware and software version identifiers from parsed MON-VER message.
 
 ### RELEASE 1.2.43